Added the null string

+++ expenses_delete.php
@@ -1,6 +1,11 @@
 <?php
 include 'db.php';
 
+if (!isset($_GET['id']) || $_GET['id'] == NULL ){
+  header("Location: signup.php");
+  exit();
+}
+
 $id = $_GET['e_id'];
 $user_id = $_GET['id'];
 $sql = "DELETE FROM expenses where e_id=$id";
